rename model names
