+++ controllers/courses_controller.js
@@ -1,15 +1,15 @@
-const getCourses = require("../models/courseDetail_model");
-const getComment = require("../models/getComment_model");
-const updateComment = require("../models/updateComment_model");
-const deleteComment = require("../models/deleteComment_model");
-const getCourseList = require("../models/courseList_model");
-const getCourse = require("../models/getCourses_model");
-const saveComment = require("../models/comment_model");
+const getCourses = require("../models/getRating_model");
+const getComment = require("../models/_getComment_model");
+const updateComment = require("../models/_updateComment_model");
+const deleteComment = require("../models/_deleteComment_model");
+const getCourseList = require("../models/_courseList_model");
+const getCourse = require("../models/_getCourses_model");
+const saveComment = require("../models/_comment_model");
 const saveLog = require("../models/saveError_model");
 const getLog = require("../models/getLog_model");
 const getUser = require("../models/getUser_model");
 const countDB = require("../models/dbCount_model");
-const checkComment = require("../models/checkComment_model");
+const checkComment = require("../models/_checkComment_model");
 const departments = require("../data/class_ids_names.json");
 
 //const roles = require("../data/roles.json");
