+++ models/_comment_model.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 var ObjectId = require("mongodb").ObjectID;
-const CourseRate = require("./rating_model");
+const CourseRate = require("./schema/rating_model");
 
 const rawdata = require(__dirname + "/../data/class_ids_names.json");
 const CourseUser = require("./user_model");
